+++ eurocropsml/acquisition/builder.py
@@ -86,6 +86,7 @@ def build_dataset(
     logger.info("Finished step 3: Clipping parcels from raster tiles.")
     region.add_nuts_regions(
         ct_config,
+        config.raw_data_dir,
         satellite_output_dir,
         shape_dir_clean,
         nuts_dir,